test_srsic.py
# Import the required modules and libraries
import unittest
import pdb
import srsic

# Define the test class for the SRSIC program
class TestSRSIC(unittest.TestCase):
    """A test class for the SRSIC program that contains the unit tests and debugging tools."""

    # Define the setup method for the test class
    def setUp(self):
        """Create a SRSIC object for testing, with the following attributes:
        - name: "srsic.py"
        - path: the current working directory
        - fitness: 0
        - data: None
        - feedback: None
        - examples: None
        """
        self.srsic = srsic.SRSIC("srsic.py", os.getcwd(), 0, None, None, None)

    # Define the test methods for the SRSIC program
    def test_replicate(self):
        """Test the replicate method of the SRSIC program, which should generate a copy of the SRSIC program in a different file and directory, with the same name and code, but a different path and fitness value."""
        # Call the replicate method of the SRSIC object
        copy = self.srsic.replicate()
        # Assert that the copy is a SRSIC object
        self.assertIsInstance(copy, srsic.SRSIC)
        # Assert that the copy has the same name and code as the original SRSIC object
        self.assertEqual(copy.name, self.srsic.name)
        self.assertEqual(copy.code, self.srsic.code)
        # Assert that the copy has a different path and fitness value than the original SRSIC object
        self.assertNotEqual(copy.path, self.srsic.path)
        self.assertNotEqual(copy.fitness, self.srsic.fitness)

    def test_modify(self):
        """Test the modify method of the SRSIC program, which should modify the code of the SRSIC program using genetic operators, and aim to improve its fitness value."""
        # Call the modify method of the SRSIC object
        self.srsic.modify()
        # Assert that the code of the SRSIC object has changed
        self.assertNotEqual(self.srsic.code, srsic.code)
        # Assert that the fitness value of the SRSIC object has increased
        self.assertGreater(self.srsic.fitness, 0)

    def test_evaluate(self):
        """Test the evaluate method of the SRSIC program, which should evaluate the performance and functionality of the SRSIC program using a fitness function, and return a numerical value."""
        # Call the evaluate method of the SRSIC object
        fitness = self.srsic.evaluate()
        # Assert that the fitness value is a numerical value
        self.assertIsInstance(fitness, (int, float))
        # Assert that the fitness value is positive
        self.assertGreaterEqual(fitness, 0)

    def test_learn(self):
        """Test the learn method of the SRSIC program, which should learn from data, feedback, and examples, using reinforcement learning and meta-learning algorithms, and aim to improve its performance and functionality on any environment and task."""
        # Call the learn method of the SRSIC object
        self.srsic.learn()
        # Assert that the SRSIC object has learned a policy network and a Q-table
        self.assertIsInstance(self.srsic.policy_network, torch.nn.Module)
        self.assertIsInstance(self.srsic.Q_table, scipy.sparse.dok_matrix)
        # Assert that the policy network and the Q-table have the correct shapes
        self.assertEqual(self.srsic.policy_network.shape, (len(self.srsic.code), 256, 256, len(self.srsic.actions)))
        self.assertEqual(self.srsic.Q_table.shape, (len(self.srsic.states), len(self.srsic.actions)))

    def test_communicate(self):
        """Test the communicate method of the SRSIC program, which should communicate and cooperate with other SRSICs, and exchange information and code, creating a collective intelligence network."""
        # Call the communicate method of the SRSIC object
        self.srsic.communicate()
        # Assert that the SRSIC object has exchanged information and code with another SRSIC object
        self.assertIsInstance(self.srsic.srsic_code_base64, bytes)
        self.assertIsInstance(self.srsic.server_code_base64, bytes)
        # Assert that the information and code are encoded as JSON and base64 strings
        self.assertEqual(self.srsic.srsic_code_base64, base64.b64encode(json.dumps(self.srsic.code).encode()))
        self.assertEqual(self.srsic.server_code_base64, base64.b64decode(json.dumps(self.srsic.server_code).encode()))

    def test_interact(self):
        """Test the interact method of the SRSIC program, which should interact with the user and the environment, and provide outputs and solutions."""
        # Call the interact method of the SRSIC object
        self.srsic.interact()
        # Assert that the SRSIC object has provided an output of the interaction
        self.assertIsInstance(self.srsic.output, str)
        # Assert that the output is meaningful and useful
        self.assertIn("The performance and functionality of the SRSIC program are:", self.srsic.output)
        self.assertIn("The code and the fitness value of the SRSIC program have been saved to a file.", self.srsic.output)

    # Define the teardown method for the test class
    def tearDown(self):
        """Delete the SRSIC object and the files and directories that were created during the testing."""
        # Delete the SRSIC object
        del self.srsic
        # Delete the files and directories that were created during the testing, using the os and shutil modules
        os.remove(os.path.join(self.srsic.path, self.srsic.name))
        shutil.rmtree(self.srsic.path)

# Run the test class using the unittest module
if __name__ == "__main__":
    unittest.main()
